Migrate the component fee to Typescript
